% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authoring.R
\name{cat_chat}
\alias{cat_chat}
\title{Outputs Chat to Markdown}
\usage{
cat_chat(x, title_levels = TRUE)
}
\arguments{
\item{x}{A Chat object.}

\item{title_levels}{logical. 제목 수준을 유지할지 여부를 결정합니다.}
}
\description{
Chat 결과를 Markdown 형식으로 출력합니다.
}
\examples{
\dontrun{
cat_chat(x)
}
}
