% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authoring.R
\name{chat2markdown}
\alias{chat2markdown}
\title{Convert Chat to Markdown}
\usage{
chat2markdown(x, title_levels = TRUE)
}
\arguments{
\item{x}{A Chat object.}

\item{title_levels}{logical. 제목 수준을 유지할지 여부를 결정합니다.}
}
\value{
character.
}
\description{
Chat 결과를 Markdown 형식으로 변환합니다.
}
\examples{
\dontrun{
chat2markdown(x)
}
}
