% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{chat_grok}
\alias{chat_grok}
\title{Chat with a xAI model}
\usage{
chat_grok(
  system_prompt = NULL,
  model = "grok-2-latest",
  api_key = Sys.getenv("XAI_API_KEY")
)
}
\arguments{
\item{system_prompt}{character. 시스템 프롬프트.}

\item{model}{character. 모델의 이름. 기본값은 "grok-2-latest".}

\item{api_key}{character. xAI API key.}
}
\value{
A Chat object.
}
\description{
xAI의 Grok API를 사용하여 모델과 대화합니다.
}
\examples{
\dontrun{
chat <- chat_grok()
chat$chat("Tell me three jokes about statisticians")
}
}
