% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authoring.R
\name{edit_openai}
\alias{edit_openai}
\title{Edit/Summary text}
\usage{
edit_openai(x, type = c("summary", "edit"))
}
\arguments{
\item{x}{A Chat object.}

\item{type}{character. 요약 또는 편집 중 하나를 선택. "summary", "edit" 중 하나.}
}
\value{
character.
}
\description{
Chat 결과를 요약하거나 편집
}
\examples{
\dontrun{
list_models_openai()
}
}
