% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authoring.R
\name{extract_subtitles}
\alias{extract_subtitles}
\title{List sub-titles in Chat}
\usage{
extract_subtitles(x, remove_subtitle = "^$")
}
\arguments{
\item{x}{A Chat object.}

\item{remove_subtitle}{character. 제거할 서브 타이틀을 지정합니다.}
}
\value{
character.
}
\description{
Chat 결과에서 서브 타이틀들을 추출
}
\examples{
\dontrun{
extract_subtitles(x)
}
}
