% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authoring.R
\name{get_docs}
\alias{get_docs}
\title{Extract documnents in Chats}
\usage{
get_docs(x = list(), subtitle = NULL, is_last = FALSE)
}
\arguments{
\item{x}{list. Chat objects.}

\item{subtitle}{character. 서브 타이틀.}

\item{is_last}{logical. 마지막 서브 타이틀 여부.}
}
\value{
character.
}
\description{
Chat 리스트에서 특정 subtitle의 내용들을 추출
}
\examples{
\dontrun{
get_docs(x, subtitle = "3. 판매채널의 유형 및 전략")
}
}
