% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{list_models_openai}
\alias{list_models_openai}
\alias{list_models_anthropic}
\title{List openai models}
\usage{
list_models_openai(api_key = Sys.getenv("OPENAI_API_KEY"))

list_models_anthropic(api_key = Sys.getenv("ANTHROPIC_API_KEY"))
}
\arguments{
\item{api_key}{character. anthropic API key.}
}
\value{
a data.frame.
Names of data.frame is as follows.
\itemize{
\item id : 모델의 고유 식별자. 이 식별자는 모델을 호출하거나 다른 API 요청에서 참조할 때 사용됨.
\item object : 객체의 종류를 나타내는 문자열. 모든 "List models" API 결과의 object 값은 "model".
\item created : 모델이 생성된 날짜와 시간.
\item owned_by : 모델의 오너.
}

a data.frame.
Names of data.frame is as follows.
\itemize{
\item type : 모델의 유형.
\item id : 모델의 고유 식별자. 이 식별자는 모델을 호출하거나 다른 API 요청에서 참조할 때 사용됨.
\item display_name : 모델의 이름.
\item created_at : 모델이 생성된 날짜와 시간.
}
}
\description{
현재 사용 가능한 모델을 나열하고 소유자 및 생성일자 등,
각 모델에 대한 기본 정보를 제공.

현재 사용 가능한 모델을 나열하고 소유자 및 생성일자 등,
각 모델에 대한 기본 정보를 제공.
}
\examples{
\dontrun{
list_models_openai()
}
\dontrun{
list_models_anthropic()
}
}
