% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{regist_google_key}
\alias{regist_google_key}
\title{Regist google API key to package file}
\usage{
regist_google_key(api_key = NULL)
}
\arguments{
\item{api_key}{character. 등록할 google API key.}
}
\description{
google과 인터페이스하기 위한 google API key를 등록합니다.
}
\details{
만약에 개인 컴퓨터가 아닌 여러 사용자가 사용하는 환경에 chat2doc 패키지를 설치한 경우라면,
API key의 보안을 위해서 regist_google_key()대신 set_google_key()를 사용하세요.
}
\examples{
\dontrun{
# 실제 사용자가 할당받은 google API key를 사용합니다.
# regist_google_key("XXXXXXXXXXX")
}
}
