% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{set_google_key}
\alias{set_google_key}
\title{Set google API key to system environment}
\usage{
set_google_key(api_key = NULL)
}
\arguments{
\item{api_key}{character. 등록할 google API key.}
}
\description{
google AI와 인터페이스하기 위한 google API key를 설정합니다.
}
\details{
만약에 여러 사용자가 사용하는 환경이 아닌 개인 컴퓨터에 chat2doc 패키지를 설치한 경우라면,
set_google_key() 대신에 매번 API key를 등록할 필요없는 regist_google_key()를 사용하세요.
}
\examples{
\dontrun{
# 실제 사용자가 할당받은 google API key를 사용합니다.
# set_google_key("XXXXXXXXXXX")
}
}
