% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trim_before_string}
\alias{trim_before_string}
\alias{remove_before_string}
\alias{remove_before_first}
\alias{extract_between}
\title{Trim before Text}
\usage{
trim_before_string(text, pattern)

remove_before_string(text, pattern)

remove_before_first(text, pattern)

extract_between(text, start_pattern, end_pattern)
}
\arguments{
\item{text}{character. original text.}

\item{pattern}{character. pattern to base.}

\item{start_pattern}{character. start pattern.}

\item{end_pattern}{character. end pattern.}
}
\value{
character.

character.

character.

character.
}
\description{
특정 문자열 및 이전의 모든 문자 제거

모든 특정 문자열 앞의 문자열 제거

첫 특정 문자열 앞의 문자열을 제거

두 문자열 패턴 사이의 문자열 추출
}
\examples{
\dontrun{
text <- "이것은 예제입니다. 삭제할 부분을 포함한 문자열입니다."
pattern <- "니"
trim_before_string(text, pattern)
}
\dontrun{
text <- "이것은 예제입니다. 삭제할 부분을 포함한 문자열입니다."
pattern <- "니"
remove_before_string(text, pattern)
}
\dontrun{
text <- "이것은 예제입니다. 삭제할 부분을 포함한 문자열입니다."
pattern <- "니"
remove_before_first(text, pattern)
}
\dontrun{
text <- "이것은 예제입니다. 삭제할 부분을 포함한 문자열입니다."
spattern <- "니"
epattern <- "삭"
extract_between(text, spattern, epattern)
}
}
