% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authoring.R
\name{common_subtitle}
\alias{common_subtitle}
\title{Common sub-titles in Chat}
\usage{
common_subtitle(x = list(), remove_subtitle = "^$")
}
\arguments{
\item{x}{list. Chat objects.}

\item{remove_subtitle}{character. 제거할 서브 타이틀을 지정합니다.}
}
\value{
character.
}
\description{
Chat 리스트에서 union 서브 타이틀들을 추출
}
\examples{
\dontrun{
common_subtitle(x)
}
}
